Return ft_atoi back to return an int

+++ libft/ft_atoi.c
@@ -6,11 +6,11 @@
 /*   By: aabounak <aabounak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/14 17:53:07 by aabounak          #+#    #+#             */
-/*   Updated: 2020/12/18 19:00:09 by aabounak         ###   ########.fr       */
+/*   Updated: 2020/12/18 19:24:20 by aabounak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long	condition(long result, long sign)
+int		condition(long result, long sign)
 {
 	if (result > 2147483648 && sign == -1)
 		return (0);
@@ -19,7 +19,7 @@ long	condition(long result, long sign)
 	return (0);
 }
 
-long	ft_atoi(const char *str)
+int		ft_atoi(const char *str)
 {
 	long	sign;
 	long	result;
